# Optimize `Log10()`