@fyfyrchik 2023-08-23
CI benchmarks agree (about the same ~20% improvement):
```
BenchmarkLog10/Log10/uint256-36                           969393              1241 ns/op               0 B/op          0 allocs/op
```

From #136 
```
BenchmarkLog10/Log10/uint256-36                     	  705776	      1597 ns/op	       0 B/op	       0 allocs/op
```